test(refactor): use separate files for unit and integration test fixtures
